requirements.txt
pydub
pyaudio
tkinter
mutagen
pillow